+++ MetalOS.Kernel/PortableExecutable.h
@@ -14,8 +14,8 @@
 class PortableExecutable
 {
 public:
-	static uintptr_t GetEntryPoint(const Handle hModule);
-	static size_t GetSizeOfImage(const Handle hModule);
+	static DWORD GetEntryPoint(const Handle hModule);
+	static DWORD GetSizeOfImage(const Handle hModule);
 	static PIMAGE_SECTION_HEADER GetPESection(const std::string& name, const uintptr_t ImageBase);
 	static PIMAGE_SECTION_HEADER GetPESection(const uint32_t index, const uintptr_t ImageBase);
 	static uintptr_t GetProcAddress(Handle hModule, const char* lpProcName);
